refactor: edit 'useAnimatedNode'
