+++ src/utils/splitNode/index.ts
@@ -1,0 +1,1 @@
+export { splitNode } from './splitNode';
